ride_project/requirements.txt
Django
djangorestframework
django-filter
psycopg2-binary
python-dotenv
pytest
pytest-django
